+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added the `utils/update-usernames` command, for ensuring all usernames match users’ email addresses, if the `useEmailAsUsername` config setting is enabled. ([#6312](https://github.com/craftcms/cms/issues/6312))
 - Added the `configure()` Twig function. ([#6731](https://github.com/craftcms/cms/pull/6731))
 - Matrix input type blocks now have an `id` field to allow modifying existing matrix blocks when using mutations with the GraphQL API.
+- Added `craft\services\ProjectConfig::DIFF_CACHE_KEY`.
 
 ### Changed
 - The GraphQL API now supports `variables` and `operationName` query string parameters. ([#6728](https://github.com/craftcms/cms/issues/6728))
@@ -15,6 +16,7 @@
 - The “System Status” setting now has “Offline” and “Online” labels.
 - Deprecation warnings can now contain Markdown formatting.
 - `craft\helpers\ArrayHelper::where()` now has a `$keepKeys` argument, which defaults to `true`.
+- `craft\helpers\ProjectConfig::diff()` now caches its results. ([#6736](https://github.com/craftcms/cms/issues/6736))
 - `craft\services\Config::setDotEnvVar()` no longer surrounds the value with quotes unless it contains a space. ([craftcms/docker#5](https://github.com/craftcms/docker/issues/5))
 
 ### Deprecated